Fix potential data loss when closing scene or string table tab
