+++ Finmer.Editor/AssetWindow.cs
@@ -7,6 +7,7 @@
  */
 
 using System;
+using System.Diagnostics;
 using Finmer.Core.Assets;
 
 namespace Finmer.Editor
@@ -28,6 +29,28 @@ namespace Finmer.Editor
             return Asset?.Name ?? String.Empty;
         }
 
+        public override void Flush()
+        {
+            base.Flush();
+
+            // Find the asset represented by this editor window
+            var new_asset = Asset;
+            var old_asset = Program.ActiveFurball.GetAssetByID(new_asset.ID);
+            Debug.Assert(old_asset != null, "Asset ID changes are not supported");
+            Debug.Assert(old_asset.Name.Equals(new_asset.Name, StringComparison.InvariantCulture), "Asset name changes must be copied by window");
+
+            // If the window still points to the same asset object, we do not need to replace anything
+            if (!ReferenceEquals(new_asset, old_asset))
+            {
+                // Otherwise, the old asset must be removed, so it can be replaced with the editor window's newly created asset
+                var assets = Program.ActiveFurball.Assets;
+                assets.Remove(old_asset);
+                assets.Add(new_asset);
+
+                // Replace the reference on the asset tree node, so we won't re-open the old asset later
+                Program.MainForm.ReplaceAssetTreeNode(old_asset, new_asset);
+            }
+        }
 
     }
